@@ -1,22 +1,32 @@
 {
   "name": "domain-variant-service",
   "version": "1.0.0",
-  "main": "src/server.js",
+  "main": "dist/server.js",
   "scripts": {
-    "start": "node src/server.js",
-    "dev": "nodemon src/server.js"
+    "build": "tsc",
+    "start": "node dist/server.js",
+    "dev": "ts-node src/server.ts",
+    "dev:watch": "nodemon --exec ts-node src/server.ts",
+    "test": "jest",
+    "clean": "rm -rf dist"
   },
   "dependencies": {
     "abort-controller": "^3.0.0",
     "cors": "^2.8.5",
+    "dotenv": "^16.4.5",
     "express": "^4.18.2",
-    "node-fetch": "^2.7.0"
+    "express-rate-limit": "^7.1.5",
+    "express-validator": "^7.0.1",
+    "node-fetch": "^2.7.0",
+    "winston": "^3.11.0"
   },
   "devDependencies": {
     "@types/cors": "^2.8.18",
     "@types/express": "^5.0.2",
+    "@types/node": "^20.11.5",
+    "@types/node-fetch": "^2.6.11",
     "nodemon": "^3.1.10",
     "ts-node": "^10.9.1",
-    "typescript": "^5.0.0"
+    "typescript": "^5.8.3"
   }
 }
